cassandra-stack/scripts/requirements.txt
psycopg2-binary==2.9.9
cassandra-driver==3.29.1
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
schedule==1.2.1
pydruid==0.6.7